+++ NEWS.md
@@ -1,4 +1,4 @@
-# boeCharts 1.2.3.9000
+# boeCharts 1.3.0
 
 ## Breaking changes